caso for usar em casa leia abaixo
criar projeto spring e depois substituir o que vier do repositorio, e ai vai conseguir "loudar".

+++ BrilhanteBackend/src/main/java/com/brilhante/brilhante/service/VendasService.java
@@ -1,6 +1,7 @@
 package com.brilhante.brilhante.service;
 
 
+import com.brilhante.brilhante.entity.Cliente;
 import com.brilhante.brilhante.entity.Vendas;
 import com.brilhante.brilhante.repository.VendasRepository;
 import java.util.List;
@@ -13,24 +14,29 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class VendasService {
     @Autowired
-    VendasRepository vendasRepository;
+    private VendasRepository vendasRepository;
+    @Autowired
+    private ClienteService clienteService;
     
-    public Long incluirVenda(Vendas vendas){
-        return vendasRepository.save(vendas).getIDVendas();
-    }
-    public Boolean excluirVenda(Long idVenda){
-    
-    try{
-        vendasRepository.deleteById(idVenda);
-        return true;
-    } catch (Exception ex){
-            System.out.println("Erro ao excluir"
-                    + "venda ID: " + idVenda
-                    + " Erro: " + ex.getLocalizedMessage());
-         return false;
+    public Long incluirVenda(Vendas vendas, Long IdCliente){
+       
+        Cliente cliente = clienteService.consultarCliente(IdCliente).get();
+        if(cliente != null){
+            vendas.setCliente(cliente);
+            return vendasRepository.save(vendas).getIDVendas();
         }
-            
+        return null;
     }
+    
+    public Boolean excluirVenda(Long idVenda){
+        
+        if(vendasRepository.getReferenceById(idVenda) != null){
+            vendasRepository.deleteById(idVenda);
+            return true;
+        }
+        return false;
+    }
+    
     public Optional<Vendas> consultarVenda(Long idVenda){
         
             return vendasRepository.findById(idVenda); 
@@ -46,10 +52,7 @@ public class VendasService {
         
         Vendas vnd = vendasRepository.getReferenceById(venda.getIDVendas());
                 if(vnd   != null){
-                    vnd.setCliente(venda.getCliente());
-                    vnd.setIDVendas(venda.getIDVendas());
                     vnd.setStatus(venda.getStatus());
-                    vnd.setVendasProduto(venda.getVendasProduto());
                    vendasRepository.save(vnd);
                  return true;
             } else {
